my_sdk_template/{{cookiecutter.project_slug}}/requirements.txt
# {{cookiecutter.project_slug}}/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
pydantic>=2.0.0
python-dotenv>=0.19.0
PyYAML>=5.4.1